LPS-72250 Refetch the model to make sure it is of the correct class
In some cases, the class expected by the staged model data handler is
different from the one returned from the persistance. For example, in
Documents & Media, the persistence works with DLFileEntry, but the
staged model data handler expects a FileEntry.

By forcing the staged model data handler to fetch the element, we'll
get a model that implements the expected interfaces.

Also note the behaviour for non grouped models. In this case, we
assume the model is company scoped, but if we find more than one model
with the same UUID we return null (which prevents the MBDiscussion
from being exported). This could be problematic if there are models
that are group scoped but don't implement GroupedModel (which
shouldn't happend in normal circumstances).