LPS-73007 Apply same workaround to Ratings
A proper solution to this requires some backwards incompatible changes
in the staging infrastructure, so it cannot be done yet due to the
restriction imposed by semantic versioning. To avoid exposing things
that will be immediatelly deprecated once that solution is merged, I'm
replicating the workaround logic and keeping it private.