LPS-86903 Isolating portlet request when running inside a template
When we use something like:

	$processor.processPortlet(
		"com.liferay.portal.kernel.servlet.taglib.ui.BreadcrumbEntry",
		$portletProviderAction.VIEW
	)

Inside a layout's tpl file the original request is used and the current
URL is stored in its WebKeys.CURRENT_PORTLET_URL attribute. The rest of
the portlets on the page will reuse that value instead of their own one
(see
https://github.com/liferay/liferay-portal/blob/bfb26d70b28ea4e9e1b19eccd49706bea72610c1/portal-kernel/src/com/liferay/portal/kernel/portlet/PortletURLUtil.java#L130-L135).

I used the same principle as in
https://github.com/liferay/liferay-portal/blob/1c0469d59597b4b2fd9622e141e068de33e5142c/util-taglib/src/com/liferay/taglib/portletext/RuntimeTag.java#L169-L174
to try to isolate the request and it seems to work.