LPS-94616 Checking permissions only on the specified folder
The mount folder and the root folder of an external repository are
slightly different things. The user might have permissions for the
former but not the latter. We are assuming too much and in repos
like sharepoint where extra authorization needs to happen when the user
actually accesses the folder this fails.