LPS-124467 Rendering with a separate component
It was a bit misleading to return a DLExternalVideo object for videos
hosted on Liferay when we only needed to get the HTML for them.

Now DLExternalVideo will only be used for actual external videos and
rendering will be performed by the DLVideoRenderer which will render
both internal and external videos without using the wrong abstractions.