LPS-127200 Avoiding undefined reference
For some reason when changing the local this input:

textarea#_com_liferay_journal_web_portlet_JournalPortlet_descriptionMapAsXMLEditor.lfr-editable

is validated but it has no form and fails.

When that happens this listener is not called and therefore the language
does not actually switch and the values are not properly saved.

There must be a better fix.