LPS-127602 PortletNamespace is sometimes not present
The name comes as:

ddm$$predefinedValue$dDmbgbcX$0$$en_US

And it does not match the regex which ends up generating a wrong name in
the form of:

ddm$$undefined$undefined$undefined$$undefined

Which in turn generates a wrong name for the editor which causes the
item selector event name not to match the one provided by the item
selector.

I am not sure if this is the source of the error or if the field should
be prefixed by the portletNamespace from the server.

The server adds the namespace here:

https://github.com/liferay/liferay-portal/blob/f84af965655ef1a88e170d031321726b36b98d73/modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-renderer/src/main/java/com/liferay/dynamic/data/mapping/form/renderer/internal/DDMFormFieldTemplateContextFactory.java#L330

However it is empty becuase it is being set here:

https://github.com/liferay/liferay-portal/blob/ecac7aa50afda3bec128538088083f9adbe29cab/modules/apps/data-engine/data-engine-taglib/src/main/java/com/liferay/data/engine/taglib/internal/servlet/taglib/util/DataLayoutTaglibUtil.java#L685

and here:

https://github.com/liferay/liferay-portal/blob/ecac7aa50afda3bec128538088083f9adbe29cab/modules/apps/data-engine/data-engine-taglib/src/main/java/com/liferay/data/engine/taglib/internal/servlet/taglib/util/DataLayoutTaglibUtil.java#L722

The final solution might be to set the right namespace there, but I do
not know.