LPS-135132 Avoiding doubly scoped configurations (company + group)
This happens when a site settings config is inheriting from a
config that has been set at the instance level.

The PID used in site settings is the one inherited from instance
settings:

  [...].scoped.[company related uuid]

When we save the configuration in site settings it gets that PID and
scopes it with a group related uuid generating a doubly scoped PID:

  [...].scoped.[company related uuid].scoped.[group related uuid]

The final config gets saved with the wrong PID and then cannot be
retrieved by:

  _configurationProvider.getGroupConfiguration(...

With these changes we are discarding the scope both to retrieve the
initial config if necessary and to generate the new PID when saving for
the first time. This way we avoid doubly scoped PIDs.