LPS-194865 Fixing multiCompany for already processed bundles
When a bundle has been processed and the server is restarted the portal
instance lifecycle listeners are not being registered.

This means the batch engine units will not be processed for new
companies created after that.

With this change we ensure the units are registered even if the
bundle has already been processed.

We also go from N listeners where N is the number of bundles to just 1
listener for all of them, which is quite good.