LPS-197353 Fixing assertions
The method testGetWorkflowTask_addWorkflowTask is not adding a new task
but reusing the existing ones that are created in the setup of the test
by the workflow instance.

This breaks the semantics of the base test becasue it assumes the add
method will actually add stuff to the database.

These tests were passing by accident before the changes to rest builder
from fa45c4418a6ed3ad9d00a3da8b5891e4ea0f06b8 because it was not taking
into account the context.