LPD-18197 The list was being copied too early
The relationship entity fields were being added to the list after
running the recursive calls. This caused the recursive calls to use an
early version of the list to build the deeper complex fields because it
was being transformed into a map inside the constructor.

Now the map is just passed by reference and every complex field using it
will have it fully populated at the end of the entity model
construction.

The entity model can now be truly circular.