LPD-49032 When requestURI contains only a languageID without a trailing slash, it should also be forwarded to the home page by using the proper friendlyURL
With LPS-2290 (9d2aa8e5e1), a '/' is considered a valid friendlyURL
and the request is then forwarded with the proper friendlyURL.

This behavior works as expected for the similar case, where a
requestURI contains only a languageId 'with' a trailing slash
(e.g. localhost:8080/en-US/). The friendlyURL is extracted from the
requestURI as '/', resulting with a forward to '/en-US/web/guest'.

The changes makes it so that a request that includes a languageId
'without' a trailing slash (e.g. localhost:8080/en-US), also
observes the same behavior to provide a consistent behavior.